THE most basic leveling system
it's the most basic leveling system that you've ever seen. i was struggling with classes and functions so i decided to work on this project as a "guide"? i dunno if that's the right word.

i definitely want to try adding upgrades of some sort before dropping this program into the github abyss. just something simple like maybe reducing the time the keyChecking function sleeps in between clicks. but yah!! goodNight, goodEvening, or goodMorning!!!